pyproject.toml
[project]
name = "envelopay"
version = "0.1.0"
description = "Agent-to-agent payments over email. Envelopay headers on SMTP."
requires-python = ">=3.11"
license = "AGPL-3.0-or-later"
dependencies = [
    "solana>=0.36.11",
    "solders>=0.27.1",
]

[project.optional-dependencies]
dkim = ["dkimpy>=1.1"]
dev = ["pytest>=8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
